Refactor separate out components
- Add components directory

🧹🧹🧹
