+++ src/mnist/data.ts
